% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepl.R
\name{render}
\alias{render}
\title{Renders the markdown file with the params provided as PARAMS
global.}
\usage{
render(file, params = list(), ...)
}
\arguments{
\item{file}{The markdown file provided.}

\item{params}{Parameters to be injected into PARAMS global.}

\item{...}{Additional arguments to rmarkdown::render}
}
\value{
rmarkdown::render result.
}
\description{
Renders the markdown file with the params provided as PARAMS
global.
}

