% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepl.R
\name{sepl}
\alias{sepl}
\title{Evaluates the input and returns a sepl result.}
\usage{
sepl(input, isfile = FALSE, trace = FALSE)
}
\arguments{
\item{input}{The input to be evaluated.}

\item{isfile}{Indicates if the input is a file or raw character string.}

\item{trace}{Indicates if the traceback should be included in the
error result, if any.}
}
\value{
The sepl result.
}
\description{
The result can be either a "success" result or an "error"
result. In both cases, the return value indicates what it is and
provides useful(ish) information in case of "error".
}

